time fixes

+++ philo/print_utils.c
@@ -6,7 +6,7 @@
 /*   By: molasz-a <molasz-a@student.42barcel>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/30 12:08:12 by molasz-a          #+#    #+#             */
-/*   Updated: 2024/06/02 18:01:34 by molasz-a         ###   ########.fr       */
+/*   Updated: 2024/06/03 15:03:18 by molasz-a         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 static size_t	ft_strlen(char *str)
 {
-	size_t	i;
+	int	i;
 
 	i = 0;
 	while (str[i])
